+++ docs/CHANGELOG.md
@@ -4,3 +4,5 @@
 - Add ST_AsOverpassBbox function
 - Add ST_OverpassDownloader function
 - Fix bug when read GeometryCollection with the ST_GeomFromGeoJSON function 
+- Allow ST_GRAPH to be executed with empty geometry
+- Allow ST_GRAPH to be executed when the input table contains an autoincrement column and not only a PK